Fully implement tokens into backend

+++ src/main/java/com/communityratesgames/rest/UserController.java
@@ -27,6 +27,19 @@ public class UserController {
     private final static Logger logger = Logger.getLogger(com.communityratesgames.rest.UserController.class);
     private UserModel userModel = new UserModel();
 
+    private Long getHeaderToken(HttpHeaders header) {
+        List<String> toklist = header.getRequestHeader("Authorization");
+        if (toklist == null || toklist.size() == 0) {
+            return null;
+        }
+
+        try {
+            return Long.parseLong(toklist.get(0));
+        } catch (NumberFormatException e) {
+            return null;
+        }
+    }
+
     @Inject
     JMSSender sender;
 
@@ -93,20 +106,36 @@ public class UserController {
     @Produces({"application/json"})
     @Consumes({"application/JSON"})
     public Response logout(@Context HttpHeaders header) {
-        List<String> toklist = header.getRequestHeader("Authorization");
-        if (toklist == null || toklist.size() == 0) {
-            return Response.status(Status.UNAUTHORIZED).entity("{\"error\":\"no auth token available in header\"}").build();
+        Long token = getHeaderToken(header);
+        if (token == null) {
+            return Response.status(Status.UNAUTHORIZED).entity("{\"error\":\"invalid auth token\"}").build();
         }
 
+        if (!AuthToken.close(token)) {
+            return Response.status(Status.UNAUTHORIZED).entity("{\"error\":\"invalid auth token\"}").build();
+        }
+        return Response.status(Status.OK).build();
+    }
+
+    @GET
+    @Path("/active")
+    @Produces({"application/JSON"})
+    public Response activeUser(@Context HttpHeaders header) {
         try {
-            Long token = Long.parseLong(toklist.get(0));
-            if (!AuthToken.close(token)) {
-                return Response.status(Status.NOT_FOUND).entity("{\"error\":\"invalid token\"}").build();
+            Long token = getHeaderToken(header);
+            if (token == null) {
+                return Response.status(Status.UNAUTHORIZED).entity("{\"error\":\"invalid auth token\"}").build();
             }
 
-            return Response.status(Status.OK).build();
-        } catch (NumberFormatException e) {
-            return Response.status(Status.BAD_REQUEST).entity("{\"error\":\"token cannot be parsed; is not a number\"}").build();
+            User u = dal.getUserToken(token);
+            if (u == null) {
+                return Response.status(Status.INTERNAL_SERVER_ERROR).entity("{\"error\":\"corrupted token in database\"}").build();
+            }
+
+            UserModel model = userModel.toModel(u, token);
+            return Response.ok(model).build();
+        } catch (PersistenceException e) {
+            return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
@@ -128,7 +157,17 @@ public class UserController {
     @Path("/update")
     @Produces({"application/JSON"})
     @Consumes({"application/JSON"})
-    public Response update(String user) {
+    public Response update(@Context HttpHeaders header, String user) {
+        Long token = getHeaderToken(header);
+        if (token == null) {
+            return Response.status(Status.UNAUTHORIZED).entity("{\"error\":\"invalid auth token\"}").build();
+        }
+
+        User u = dal.getUserToken(token);
+        if (u.getRole() != "Admin") {
+            return Response.status(Status.FORBIDDEN).entity("{\"error\":\"/update may only be used by admin users\"}").build();
+        }
+
         try {
             User um = userModel.toEntity(user, false);
             System.out.println(um);
@@ -145,8 +184,18 @@ public class UserController {
     @Path("/delete")
     @Produces({"application/json"})
     @Consumes({"application/JSON"})
-    public Response delete(String username) {
+    public Response delete(@Context HttpHeaders header, String username) {
+        Long token = getHeaderToken(header);
+        if (token == null) {
+            return Response.status(Status.UNAUTHORIZED).entity("{\"error\":\"invalid auth token\"}").build();
+        }
+
         try {
+            User u = dal.getUserToken(token);
+            if (u.getRole() != "Admin") {
+                return Response.status(Status.FORBIDDEN).entity("{\"error\":\"/delete may only be used by admin users\"}").build();
+            }
+
             User um = userModel.toEntity(username, false);
             Boolean answer = dal.deleteAUser(um);
             if ( answer == true ) {
